worker: submit cancel request
Nobody will, as far as I can tell, submit past this point. So do the
submit inline from worker_iou_stop_test().

Also, since the request is now being submitted and processed, add a
command type for cancel requestion. Since it's just for exit, nothing
is being done with it. But it avoids an error ala:

server: Unknown io_uring request type: 0, res: -2: Address already in use

when the request is reaped.

Signed-off-by: Jens Axboe <axboe@kernel.dk>